grid_bot/requirements.txt
ccxt>=4.0.0
python-telegram-bot>=20.0
pydantic>=2.0.0
